% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSmoothFD.R
\name{createSmoothFD}
\alias{createSmoothFD}
\title{Create a smooth functional data object using bsplines}
\usage{
createSmoothFD(curvesToSmooth, abscissa, lambdas = exp(-5:5),
  basisBreakFreq = 3, basisOrder = 4, pbasis = NULL)
}
\arguments{
\item{curvesToSmooth}{A matrix or a vector to smooth}

\item{abscissa}{The x-axis. Should of of the same length as the vector curvesToSmooth or the 
number of rows in matrix curvesToSmooth}

\item{lambdas}{The different roughness penalty to try during smoothing, for the parameter 
\code{lambda} in \code{fda:fdPar}}

\item{basisBreakFreq}{Defaults to 3. More frequent could overfit, less frequent could over-smooth.}

\item{basisOrder}{Defaults to 4. If more noisy, then this can be higher.}

\item{pbasis}{User-defined \code{pbasis} if user-defined basisObject is desired. Else defaults to NULL.}
}
\value{
A list of two items
\item{curvesSmooth}{FDA object of smoothed curves}
\item{lambdaBest}{lambda associated with minimum GCV}
}
\description{
For a curve of a matrix this function creates a smooth functional data object as per 
the definition in R package fda, using bspline basis.
}
\examples{
data( growth, package = 'fda' )
Mat1 <- growth[['hgtm']]
Mat1Smooth <- createSmoothFD( 
   curvesToSmooth = Mat1, 
   abscissa = 1:nrow( Mat1 ), 
   lambdas = exp( -5:5 ), 
   basisBreakFreq = 3, 
   basisOrder = 4, 
   pbasis = NULL )
#plot( (Mat1Smooth$curvesSmooth)$fd )

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\references{
Graves, S., Hooker, G., & Ramsay, J. (2009). Functional data analysis with R and MATLAB.
}
\seealso{
\code{fda::create.basis}, \code{fda::create.bspline.basis}
}

