% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeSmoothFunction.R
\name{encodeSmoothFunction}
\alias{encodeSmoothFunction}
\title{Encodes a smooth function into a sequence of states}
\usage{
encodeSmoothFunction(smoothFunc, abscissa)
}
\arguments{
\item{smoothFunc}{A smooth function, already converted into an \code{fda::fd} object.}

\item{abscissa}{A vector representing the x-axis or abscissa}
}
\value{
A \code{string} states representing the encoded form of the smooth function, example "AAHHFFFGGH"
}
\description{
Encodes a smooth function by its values, first and second derivatives. At each point on
the abscissa the function is encoded into a three dimensional point. These points are represented as
sequence of states
}
\details{
To use this function for a vector use the function \code{createSmoothFD} to convert to a 
\code{fd} object.
For a smooth function \eqn{f(x)}, the eight possible states are as follows:
\itemize{
   \item \eqn{A=(1,1,1): (f>0, f'>0, f''>0)}
   \item \eqn{B=(1,0,1): (f>0, f'\leq 0, f''>0)}
   \item \eqn{C=(1,1,0): (f>0, f'> 0, f''\leq 0)}
   \item \eqn{D=(1,0,0): (f>0, f'\leq 0, f''\leq 0)}
   \item \eqn{E=(0,1,1): (f\leq 0, f'> 0, f''>0)}
   \item \eqn{G=(0,0,1): (f\leq 0, f'\leq 0, f''> 0)}
   \item \eqn{H=(0,1,0): (f\leq 0, f'> 0, f''\leq 0)}
   \item \eqn{I=(0,0,0): (f\leq 0, f'\leq 0, f''\leq 0)}
}
}
\examples{
 data( growth, package = 'fda')
 Mat1 <- growth[['hgtm']]
 Arguments <- growth[['age']]
 Mean <- rowMeans( Mat1 )
 MeanFD <- createSmoothFD( 
     curvesToSmooth = Mean,
     abscissa       = Arguments,
     lambdas        = exp( -5:5 ),
     basisBreakFreq = 3,
     basisOrder     = 4,
     pbasis         = NULL )
 MeanStates <- encodeSmoothFunction( smoothFunc = (MeanFD$curvesSmooth)$fd, abscissa = Arguments )

}
\author{
Subhrangshu Nandi; snandi@wisc.edu or nands31@gmail.com
}

