% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeSmoothFunctionOrder1.R
\name{encodeSmoothFunctionOrder1}
\alias{encodeSmoothFunctionOrder1}
\title{Encodes a smooth function into a sequence of states}
\usage{
encodeSmoothFunctionOrder1(smoothFunc, abscissa)
}
\arguments{
\item{smoothFunc}{A smooth function, already converted into an \code{fda::fd} object.}

\item{abscissa}{A vector representing the x-axis or abscissa}
}
\value{
A \code{string} of states representing the encoded form of the smooth function, example "AADDCCCB"
}
\description{
Encodes a smooth function by its first derivatives. At each point on
the abscissa the function is encoded into a binary variable. These points are represented as
sequence of states
}
\details{
To use this function for a vector use the function \code{createSmoothFD} to convert to a 
\code{fd} object.
For a smooth function \eqn{f(x)}, the two possible states are as follows:
\itemize{
   \item \eqn{A=(1): (f'>0)}
   \item \eqn{B=(0): (f'\leq 0)}
}
}
\examples{
 data( growth, package = 'fda')
 Mat1 <- growth[['hgtm']]
 Arguments <- growth[['age']]
 Mean <- rowMeans( Mat1 )
MeanFD <- createSmoothFD(
    curvesToSmooth = Mean,
    abscissa       = Arguments,
    lambdas        = exp( -5:5 ),
    basisBreakFreq = 3,
    basisOrder     = 4,
    pbasis         = NULL )
 MeanStates <- encodeSmoothFunctionOrder1( smoothFunc = (MeanFD$curvesSmooth)$fd, abscissa = Arguments )

}
\author{
Subhrangshu Nandi, PhD Statistics; snandi@wisc.edu or nands31@gmail.com
}

