% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunctionalOutliers.R
\name{getFunctionalOutliers}
\alias{getFunctionalOutliers}
\title{Gets functional outliers from a sample of curves}
\usage{
getFunctionalOutliers(Curves, Xaxis, Names = list(main = "Main", xlab =
  "PixelPosition", ylab = "Intensity"), DepthType = c("FM", "Mode", "RTukey",
  "RProj"), N_Bootstrap = 500, Trim = c("Yes", "No"), TrimPct = 0.05)
}
\arguments{
\item{Xaxis}{The abscissa}

\item{Names}{list(main = 'Main', xlab = 'PixelPosition', ylab = 'Intensity')}

\item{DepthType}{Either of the four \code{c( 'FM', 'Mode', 'RTukey', 'RProj' )}}

\item{N_Bootstrap}{Number of boostrap samples. Defaults to 500.}

\item{Trim}{Whether to Trim the samples or not.}

\item{TrimPct}{How much to Trim the samples or not}

\item{curves}{A matrix (or dataframe) of curves, with each column being a separate curve}
}
\value{
A vector of column names that are detected as outliers.
}
\description{
Gets functional outliers from a sample of curves using methods described in fda.usc
}
\examples{
data( growth, package = 'fda' )
Mat1 <- growth[['hgtm']]
Arguments <- growth[['age']]
getFunctionalOutliers( 
  Curves = Mat1, 
  Xaxis = Arguments, 
  Names = list(main = 'Main', xlab = 'PixelPosition', ylab = 'Intensity'),
  DepthType = 'FM',
  N_Bootstrap = 500,
  Trim = 'Yes',
  TrimPct = 0.05
)

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}
\references{
\itemize{
\item Febrero-Bande, M., & Oviedo de la Fuente, M. (2012). Statistical computing in functional data analysis: the R package fda. usc. Journal of Statistical Software, 51(4), 1-28.
\item Cuevas A, Febrero M, Fraiman R. 2006. On the use of bootstrap for estimating functions with functional data. Computational Statistics and Data Analysis 51: 1063-1074.
\item Febrero-Bande, M., Galeano, P., and Gonzalez-Manteiga, W. (2008). Outlier detection in functional data by depth measures with application to identify abnormal NOx levels. Environmetrics 19, 4, 331-345.
\item Febrero-Bande, M., Galeano, P. and Gonzalez-Manteiga, W. (2007). A functional analysis of NOx levels: location and scale estimation and outlier detection. Computational Statistics 22, 3, 411-427.
\item Febrero-Bande, M., Oviedo de la Fuente, M. (2012). Statistical Computing in Functional Data Analysis: The R Package fda.usc. Journal of Statistical Software, 51(4), 1-28. 
\item http://www.jstatsoft.org/v51/i04/
}
}
\seealso{
\code{outliers.depth.trim}, \code{outliers.depth.pond}
}
\keyword{data}
\keyword{depth}
\keyword{functional}
\keyword{outliers}

