% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKmaSimilarityMatrix.R
\name{getKmaSimilarityMatrix}
\alias{getKmaSimilarityMatrix}
\title{Get a pairwise similarity matrix of a group of curves}
\usage{
getKmaSimilarityMatrix( Mat, Xaxis, similarity.method = c("d1.pearson", "d1.L2", "d1.L2.centered") )
}
\arguments{
\item{Mat}{Matrix with each column representing a separate curve}

\item{Xaxis}{Vector of numbers, the abscissa}

\item{similarity.method}{Type of similarity method to use}
}
\value{
List of two objects
\item{ColCombs}{A data frame of pairwise comparison similarity index}
\item{AvgSimilarity}{Average of all similarities}
}
\description{
Returns pairwise similarity between columns of a matrix of first derivatives of curves, 
and the average of the similarities
}
\details{
Assumed that the x-axis of all the functions will be the same 
d1 corresponds to first derivatives. Make sure the matrix Mat is that of
first derivatives and not just the function values
}
\examples{
data( growth, package = 'fda')
Mat1 <- growth[['hgtm']]
Arguments <- growth[['age']]
getKmaSimilarityMatrix( Mat = Mat1, Xaxis = Arguments, similarity.method = "d1.pearson" )

}
\author{
Subhrangshu Nandi; Statistics PhD student, UW Madison; snandi@wisc.edu or nands31@gmail.com
}

