% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqCompCount.R
\name{getSeqCompCount}
\alias{getSeqCompCount}
\title{A dataframe of counts of k-mers}
\usage{
getSeqCompCount(seqData, maxNmers = 5, seqWindow = 206)
}
\arguments{
\item{seqData}{Sequence data of class \code{SeqFastadna}}

\item{maxNmers}{An integer between 1 and 5}

\item{seqWindow}{Defaults to 206. This is the subsequence window in which the counts of different 
k-mers is returned}
}
\value{
A dataframe of counts of k-mers. Each column corresponds to a k-mer, and each row corresponds
to a 206 \code{seqWindow}
}
\description{
Return a dataframe of counts of k-mers of a user provided genomic sequence
}
\examples{
fpath <- system.file( "extdata", "chr22_Segment750.fa", package = "FscanStats" )
seqData <- seqinr::read.fasta( file = fpath )[[1]]
segmentSeqCompData <- getSeqCompCount( 
   seqData   = seqData,
   maxNmers  = 4
)

}
\author{
Subhrangshu Nandi, PhD Statistics, UW Madison; snandi@wisc.edu or nands31@gmail.com
}

