% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubsequenceChunks.R
\name{getSubsequenceChunks}
\alias{getSubsequenceChunks}
\title{Get subsequence chunks from a long sequence}
\usage{
getSubsequenceChunks(Sequence, numChunks, force.number.of.groups = TRUE)
}
\arguments{
\item{Sequence}{A DNA sequence in the form of a list of characters: "a" "g" "g" "c" "t", similar 
to the output of \code{seqinr::s2c}}

\item{numChunks}{Number of subsequence chunks desired}

\item{force.number.of.groups}{Defaults to \code{TRUE}. If \code{TRUE}, then it will append the 
remainder of the elements to the last chunk}
}
\value{
Returns a list of subsequence chunks.
}
\description{
This function takes in a DNA sequence "agggctattaaccctttaaa" in the form of a list of 
characters and returns smaller subsequence chunks based on user specified number.
}
\examples{
getSubsequenceChunks( 
  Sequence = "agggctattaaccctttaaa", 
  numChunks = 3, 
  force.number.of.groups = TRUE )

}
\author{
Subhrangshu Nandi; PhD, Statistics; snandi@wisc.edu or nands31@gmail.com
}
\keyword{DNA-sequence}
\keyword{subsequence}

