% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.rigid.R
\name{logLik.translate.rotate2d.Matern.allp}
\alias{logLik.translate.rotate2d.Matern.allp}
\title{Log likelihood function with embedded registration}
\usage{
\method{logLik}{translate.rotate2d.Matern.allp}(p, nu, grd, grd2)
}
\arguments{
\item{p}{parameters defining Gaussian process covariance and registration}

\item{nu}{smoothness parameter of GP}

\item{grd}{three column matrix of point set 1}

\item{grd2}{three column matrix of point set 2}
}
\value{
the value of the negative log likelihood
}
\description{
Registration includes translations in the x, y, and z coordinates and a rotate in the xy plane
}
