% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseVariability.R
\name{phaseVariability}
\alias{phaseVariability}
\title{Phase variability of a sample of curves}
\usage{
phaseVariability(curves, template = NULL)
}
\arguments{
\item{curves}{A \code{matrix} or \code{data.frame} of curves, each column representing a separate curve}

\item{template}{A template curve. If \code{NULL} the template is estimated as \code{rowMeans(curves)}}
}
\value{
phaseVariability \eqn{\omega}, where \eqn{0 \leq \omega \leq 1}
}
\description{
This function estimates the phase variability of a sample of curves that are considered
random realizations of a smooth function
}
\details{
This is the first mathematical definition of phase variability. It uses similarity index 
(\eqn{\rho}), defined in \code{fdakma::kma.similarity}. For a sample of \eqn{n} curves that are random 
realizations of a smooth template curve, phase variability is defined as 
\eqn{\omega = \frac{1 - \rho_m}{2}}, where \eqn{\rho_m = \frac{1}{n}\sum\rho_i}, \eqn{\rho_i} being the 
similarity between curve \eqn{i} and the template.
}
\examples{
data( growth, package = 'fda' )
Mat1 <- growth[['hgtm']]
phaseVar <- phaseVariability( curves = Mat1 )

}
\author{
Subhrangshu Nandi; PhD, Statistics; snandi@wisc.edu or nands31@gmail.com
}
\keyword{phase-variability}
\keyword{similarity}

