% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmseCurves.R
\name{rmseCurves}
\alias{rmseCurves}
\title{RMSE of a sample of curves with a template}
\usage{
rmseCurves(curves, template = NULL, Round = 4)
}
\arguments{
\item{curves}{A \code{matrix} or \code{data.frame} of curves, each column being a separate curve}

\item{template}{A template curve. If \code{NULL} the template is estimated as \code{rowMeans(curves)}}

\item{Round}{An integer value to round off the average rmse.}
}
\value{
Returns a list
\item{rmseVector}{A vector of rmse between curves and template}
\item{rmseMean}{Average of \code{rmseVector}}
}
\description{
This function takes in a sample of curves in the form of a \code{matrix} or \code{data.frame}
with each column being a separate curve, and returns the average RMSE (root mean squared error) between
the curves and a template.
}
\examples{
data( growth, package = 'fda' )
Mat1 <- growth[['hgtm']]
rmseMat1 <- rmseCurves( curves = Mat1 )

}
\author{
Subhrangshu Nandi; PhD, Statistics; snandi@wisc.edu or nands31@gmail.com
}
\keyword{rmse}

