% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.rigid.R
\name{rotate2d}
\alias{rotate2d}
\title{rotate xy coordinates}
\usage{
rotate2d(df, rt)
}
\arguments{
\item{df}{the dataframe with coordinates in first three columns}

\item{rt}{a rotation angle in radians}
}
\value{
the dataframe with rotated coordinates
}
\description{
rotate xy coordinates
}
