% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SeqCompData}
\alias{SeqCompData}
\title{Sequence composition data}
\format{A dataframe This data has columns with sequence compositions. Each row corresponds to 206bp of genomic DNA.}
\description{
This data has columns with sequence compositions. Each row corresponds to 206bp of genomic DNA.
}
\keyword{datasets}

