% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.rigid.R
\name{translate}
\alias{translate}
\title{Translate the x and y-coordinates in the first three columns}
\usage{
translate(df, tr)
}
\arguments{
\item{df}{the dataframe with coordinates in first three columns}

\item{tr}{a vector of length 3 with the x, y, and z translations}
}
\value{
the dataframe with translated coordinates
}
\description{
Translate the x and y-coordinates in the first three columns
}
