% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeSmoothFunction.R
\name{encodeSmoothFunction}
\alias{encodeSmoothFunction}
\title{Encodes a smooth function into a sequence of states}
\usage{
encodeSmoothFunction(smoothFunc, abscissa)
}
\arguments{
\item{smoothFunc}{A smooth function, already converted into an \code{fda::fd} object.}

\item{abscissa}{A vector representing the x-axis or abscissa}
}
\value{
A \code{string} of encoded form of the function in the form of a sequence of states.
}
\description{
Encodes a smooth function by its values, first and second derivatives. At each point on
the abscissa the function is encoded into a three dimensional point. These points are represented as
sequence of states
}
\details{
To use this function for a vector use the function \code{createSmoothFD} to convert to a 
\code{fd} object.
}
\examples{
 data( growth, package = 'fda')
 Mat1 <- growth[['hgtm']]
 Arguments <- growth[['age']]
 Mean <- rowMeans( Mat1 )
 MeanFD <- createSmoothFD( 
 curvesToSmooth = Mean,
 abscissa       = Arguments,
 lambdas        = exp( -5:5 ),
 basisBreakFreq = 3,
 basisOrder     = 4,
 pbasis         = NULL )
 MeanStates <- encodeSmoothFunction( smoothFunc = (MeanFD$curvesSmooth)$fd, abscissa = Arguments )

}
\author{
Subhrangshu Nandi; snandi@wisc.edu or nands31@gmail.com
}

